Power BI Churn Analysis
I have joined a virtual internship of PwC through forage, where my role was of Power BI Analyst. This was the 2nd task assigned in internship, objective was that the client wanted insights behind the churning customers.
My first approach was to check the data & perform cleaning for better results & started my analysis on Power BI Desktop.

Insights:- 

➢ Customer Overview:
There is total 7,043 customers subscribed to the client, out of which 3,555 are Males & 3,488 are Females, & the revenue earned is $16.1M
84% of our customers are Young i.e. 5901.
By tenure, 31% customers are subscribed to below 12 months, as a result there are more customers with month-to-month contract.
Most customers have Fiber optic internet for which they are paying high charges.
73.5% of customers have been retained i.e. 5,174.

➢ Churn Customers:
1,869 customers have churned i.e. 26.5%, out of which 380 customers churned last month itself.
Revenue lost due to churn is $2.9M.
Gender is not the major factor of churning as ratio of Males and Females is same.
1037 customers with less than 12 months tenure have churned, whereas customers with 61-72 months of tenure are just 93.
89% customers churned had month-to-month contract.
Customers paying high charges have churned the most i.e.1274, along with customers with fiber optic internet have also churned the most

➢ Churning factors are:
Young Customers
Customers with tenure of last than 12 months
Customers with month-to-month contract
Customers subscribed to fiber optic.
Customers paying high charges.